requirements.txt
tensorflow>=2.12.0
pillow>=9.5.0
tensorflow_datasets>=4.9.2
tqdm==4.65.0